Add arrows to transaction list
