+++ apps/wallet/src/app/shared/components/index.ts
@@ -3,5 +3,4 @@ export * from './credentials-form';
 export * from './copyable-text';
 export * from './loading-spinner';
 export * from './transaction-details';
-export * from './transaction-row';
 export * from './transactions-list';
