--- apps/wallet/src/app/shared/components/transaction-row/transaction-row.component.html
@@ -1,3 +1,0 @@
-<p>
-  transaction-row works!
-</p>
