--- apps/wallet/src/app/shared/components/transaction-row/transaction-row.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'lto-transaction-row',
-  templateUrl: './transaction-row.component.html',
-  styleUrls: ['./transaction-row.component.scss']
-})
-export class TransactionRowComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
