--- apps/wallet/src/app/shared/components/transaction-row/transaction-row.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { TransactionRowComponent } from './transaction-row.component';
-
-@NgModule({
-  declarations: [TransactionRowComponent],
-  imports: [CommonModule],
-  exports: [TransactionRowComponent]
-})
-export class TransactionRowModule {}
