+++ apps/wallet/src/app/shared/components/transactions-list/transaction-row/index.ts
@@ -1,2 +1,1 @@
 export * from './transaction-row.component';
-export * from './transaction-row.module';
