+++ apps/wallet/src/app/shared/components/transactions-list/transactions-list.component.html
@@ -2,54 +2,64 @@
   <div class="group" *ngFor="let group of transactions; trackBy: groupDayTrackBy">
     <div class="date">{{ group.date }}</div>
     <div class="transactions">
-      <div
-        class="transaction"
+      <lto-transaction-row
         *ngFor="let transaction of group.transactions; trackBy: transactionTrackBy"
-        fxLayout="row"
-        fxLayoutAlign="start center"
+        [transaction]="transaction"
+        [myAddress]="address"
         (click)="_transactionClick(transaction)"
         [ngClass]="{ selected: selectedId === transaction.id }"
-      >
-        <div class="info">
-          <div class="addresses">
-            <ng-container *ngIf="address === transaction.sender">
-              <span class="label">TO:</span> <span class="value">{{ transaction.recipient }}</span>
-            </ng-container>
+      ></lto-transaction-row>
+      <!--
+        <div
+          class="transaction"
+          *ngFor="let transaction of group.transactions; trackBy: transactionTrackBy"
+          fxLayout="row"
+          fxLayoutAlign="start center"
+          (click)="_transactionClick(transaction)"
+          [ngClass]="{ selected: selectedId === transaction.id }"
+        >
+          <div class="icon"><mat-icon>arrow_downward</mat-icon></div>
+          <div class="info">
+            <div class="addresses">
+              <ng-container *ngIf="address === transaction.sender">
+                <span class="label">TO:</span> <span class="value">{{ transaction.recipient }}</span>
+              </ng-container>
 
-            <ng-container *ngIf="address !== transaction.sender">
-              <span class="label">FROM:</span> <span class="value">{{ transaction.sender }}</span>
-            </ng-container>
+              <ng-container *ngIf="address !== transaction.sender">
+                <span class="label">FROM:</span> <span class="value">{{ transaction.sender }}</span>
+              </ng-container>
+            </div>
+            <div class="misc" fxLayout="row" fxLayoutAlign="space-between">
+              <div class="time">
+                <span class="label">TIME:</span>
+                <span class="value">{{ transaction.timestamp | date: 'HH:MM' }}</span>
+              </div>
+              <div class="type">
+                <span class="label">TYPE:</span>
+                <span class="value">{{ transaction.type | typeLabel }}</span>
+              </div>
+              <div class="fee">
+                <span class="label">FEE:</span>
+                <span class="value">{{ transaction.fee | amountDivide }}</span>
+              </div>
+            </div>
           </div>
-          <div class="misc" fxLayout="row" fxLayoutAlign="space-between">
-            <div class="time">
-              <span class="label">TIME:</span>
-              <span class="value">{{ transaction.timestamp | date: 'HH:MM' }}</span>
+          <span fxFlex="1 1 auto"></span>
+          <div class="amount" fxLayout="column" fxLayoutAlign="start center">
+            <div>
+              <span class="label">{{ transaction.amount | amountDivide }}</span> <span>LTO</span>
             </div>
-            <div class="type">
-              <span class="label">TYPE:</span>
-              <span class="value">{{ transaction.type | typeLabel }}</span>
-            </div>
-            <div class="fee">
-              <span class="label">FEE:</span>
-              <span class="value">{{ transaction.fee | amountDivide }}</span>
-            </div>
+            <button
+              *ngIf="isLease(transaction)"
+              mat-button
+              color="warn"
+              (click)="_cancelLease(transaction, $event)"
+            >
+              Cancel lease
+            </button>
           </div>
         </div>
-        <span fxFlex="1 1 auto"></span>
-        <div class="amount" fxLayout="column" fxLayoutAlign="start center">
-          <div>
-            <span class="label">{{ transaction.amount | amountDivide }}</span> <span>LTO</span>
-          </div>
-          <button
-            *ngIf="isLease(transaction)"
-            mat-button
-            color="warn"
-            (click)="_cancelLease(transaction, $event)"
-          >
-            Cancel lease
-          </button>
-        </div>
-      </div>
+      -->
     </div>
   </div>
 </div>
