+++ apps/wallet/src/app/shared/components/transactions-list/transactions-list.module.ts
@@ -3,16 +3,18 @@ import { CommonModule } from '@angular/common';
 import { TransactionsListComponent } from './transactions-list.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { AmountDividePipeModule, TypeLabelPipeModule } from '../../pipes';
-import { MatButtonModule } from '@angular/material';
+import { MatButtonModule, MatIconModule } from '@angular/material';
+import { TransactionRowComponent } from './transaction-row';
 
 @NgModule({
-  declarations: [TransactionsListComponent],
+  declarations: [TransactionsListComponent, TransactionRowComponent],
   imports: [
     CommonModule,
     FlexLayoutModule,
     AmountDividePipeModule,
     TypeLabelPipeModule,
-    MatButtonModule
+    MatButtonModule,
+    MatIconModule
   ],
   exports: [TransactionsListComponent]
 })
